Improve delete photo function
If no other user is uploaded the same photo, delete the photo from the file system. Also if the photo is being used as the current profile picture, set profile_pic_path to NULL in the users table (Could revert to a previous profile pic, but we choose to let the user decide for herself).